keep track of sessions being opened so don't open twice if user double clicks in onbording window. Fixed target ref in welcome.html so uses the wspace element, not element that was actually clicked (like italic tag).

+++ MacClient/OnboardingViewController.swift
@@ -78,11 +78,11 @@ extension OnboardingViewController: WKScriptMessageHandler {
 			openLocalWorkspace?(nil)
 			break
 		case "open":
-			guard let projId = project?.projectId, let wspaceId = args["wspaceId"] as? Int else {
+			guard let projId = project?.projectId, let wspaceId = args["wspaceId"] as? Double else {
 				os_log("invalid params to open a session", log: .app, type: .error)
 				return
 			}
-			openLocalWorkspace?(WorkspaceIdentifier(projectId: projId, wspaceId: wspaceId))
+			openLocalWorkspace?(WorkspaceIdentifier(projectId: projId, wspaceId: Int(wspaceId)))
 		default:
 			os_log("unhandled action: %{public}s", log: .app, action)
 		}
